requirements.txt
absl-py==1.0.0
anyio==3.3.4
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
gast==0.3.3
google-pasta==0.2.0
grpcio==1.41.1
h5py==2.10.0
idna==2.10
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
joblib==1.2.0
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyter-server==1.11.2
jupyterlab==3.2.3
jupyterlab-language-pack-zh-CN==3.2.post0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lightgbm==3.3.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.4
nbclient==0.5.5
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.18.5
nvidia-cublas==11.5.1.101
nvidia-cublas-cu114==11.6.5.2
nvidia-cuda-cupti-cu114==11.4.120
nvidia-cuda-nvcc-cu114==11.4.120
nvidia-cuda-nvrtc==11.3.58
nvidia-cuda-runtime==11.3.58
nvidia-cuda-runtime-cu114==11.4.108
nvidia-cudnn==8.2.0.51
nvidia-cudnn-cu114==8.2.4.15
nvidia-cufft-cu114==10.5.2.100
nvidia-curand-cu114==10.2.5.120
nvidia-cusolver-cu114==11.2.0.120
nvidia-cusparse-cu114==11.6.0.120
nvidia-dali-cuda110==1.6.0
nvidia-dali-nvtf-plugin==1.6.0+nv21.10
nvidia-nccl-cu114==2.11.4
nvidia-pyindex==1.0.9
nvidia-tensorflow==1.15.5+nv21.10
nvidia-tensorrt==8.0.3.4
opt-einsum==3.3.0
packaging==21.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.1.3
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pycosat==0.6.3
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.25.1
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
setuptools==52.0.0
six==1.16.0
sniffio==1.2.0
supervisor==4.2.2
tensorboard==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.61.2
traitlets==5.1.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
wheel==0.36.2
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.6.0
